requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# LangChain & RAG
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.43
langchain-text-splitters==0.2.4

# vector db
chromadb==0.5.5
pypdf==4.0.1


# Embeddings & LLM
sentence-transformers==3.0.1
torch==2.2.2
transformers==4.37.0




# google gemini
langchain-google-genai==1.0.10
google-generativeai==0.7.2

# ML & Clustering
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.2.0

# MLflow
mlflow==2.9.2

# Utils
python-dotenv==1.0.1
pyyaml==6.0.1
requests==2.32.3

argon2-cff